Add files via upload
This is Korea Univ. 2021 data structure course's first homework.
codes except for 'main' function are made by me

the codes read txt files(which have people's names and number of the people who are born in the relevant year)
and by name, the numbers born in each year are arranged in a dynamic array.